pyproject.toml
[project]
name = "langchain-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.1.0",
    "faiss-cpu>=1.7.4",
    "openai>=1.0.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "pytest>=8.0.0",
    "tenacity>=9.1.2",
    "langchain-text-splitters>=1.1.0",
    "watchdog>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["markdown_qa"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest-asyncio>=1.3.0",
    "ty>=0.0.12",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
